Move coremod classes
