+++ gradle.properties
@@ -20,4 +20,4 @@ use_access_transformer = true
 
 # Coremod Arguments
 include_mod = true
-coremod_plugin_class_name = org.dimdev.jeid.JEIDLoadingPlugin
+coremod_plugin_class_name = org.dimdev.jeid.core.JEIDLoadingPlugin
